[Form] Precalculated the closure for deciding whether a choice is selected (PHP +30ms, Twig +30ms)
